Lots of work done
A lot was done in the pages between last commit and this one, so It would be better to just list what I have left to add to the pages.
What's left to add:
Keywords for the social media and contact pages,
Figure out the exact links for social media's that have dedicated pages for a user's account,
Figure out how to space out the distance between the header navigation, preferably without just inserting spaces at the end of the displayed text for the link.