Keywords
Added the keywords for the contact and social media pages, and added a keyword to the index/home page.
What remains to be done:
Space out the header navigation,
Add the urls for links that lead the user to other sites,
If possible, find pictures for at least two of the different social media sites I use that are listed under creative commons, and add them to the social media page by the social media they correspond with.