div victory
changed all the section tags that were applying the css from the external css sheet to div tags in order to get rid of warnings, also got the validator to pass all files. Now just the fonts need to be added.
Remember to use font awesome for icons/logos as well.